# docs/swagger.yaml
basePath: /api
definitions:
  models.Pomodoro:
    properties:
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      title:
        type: string
      todo:
        $ref: '#/definitions/models.Todo'
      todo_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Todo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      progress:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      pomodoros:
        items:
          $ref: '#/definitions/models.Pomodoro'
        type: array
      provider:
        type: string
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: api.tascaid.com
info:
  contact: {}
  description: API untuk aplikasi Tasca A4 AGILE menggunakan gin
  title: Tag Tasca API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a token
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'token: JWT Token'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /api/logout:
    post:
      description: Remove authentication token
      responses:
        "200":
          description: 'message: Logout berhasil'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
...
  /api/profile:
    get:
      description: Retrieve user profile data
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: User tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - profile
  /api/profile/change-password:
    post:
      consumes:
      - application/json
      description: Allow authenticated user to change password
      parameters:
      - description: Password Update
        in: body
        name: password
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Password berhasil diubah'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad request'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - profile
  /api/profile/delete:
    delete:
      description: Delete an authenticated user account
      responses:
        "200":
          description: 'message: Akun berhasil dihapus'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad request'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user account
      tags:
      - profile
  /api/profile/delete-picture:
    delete:
      description: Remove user's profile picture and set to default
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Foto profil berhasil dihapus, picture: nopp.png'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: User tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal menghapus foto profil'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user profile picture
      tags:
      - profile
  /api/profile/update:
    post:
      consumes:
      - multipart/form-data
      description: Upload and update user's profile
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Foto profil berhasil diperbarui,: filename'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: File tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: User tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal menyimpan file'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - profile
  /api/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Registrasi berhasil'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register user
      tags:
      - auth
...
swagger: "2.0"